requirements.txt
Django==1.8.3
djangorestframework==3.1.3
httpie==0.9.2
psycopg2==2.6.1
pymongo==3.0.3
requests==2.7.0
